Almost completed the XMLreader.getAccolades function. Added extra functions as required. Also added a poorly constructed example accolade.xml file; i plan to addd a few more with different contexts

+++ deco2800.arcade.server/src/main/java/deco2800/server/database/AccoladeStorage.java
@@ -129,7 +129,7 @@ public class AccoladeStorage{
 									resultSet.getString("description"),
 									resultSet.getInt("popup"), 
 									resultSet.getString("popupMessage"),
-									resultSet.getFloat("modifier"),
+									resultSet.getDouble("modifier"),
 									resultSet.getString("unit"),
 									resultSet.getString("tag"),
 									resultSet.getString("imagepath")
